// web/src/api/model/reservationStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Vartalaap API
 * Voice bot platform for local Indian businesses
 * OpenAPI spec version: 0.1.0
 */

/**
 * Status of a reservation.
 */
export type ReservationStatus =
  (typeof ReservationStatus)[keyof typeof ReservationStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ReservationStatus = {
  confirmed: "confirmed",
  cancelled: "cancelled",
  completed: "completed",
  no_show: "no_show",
} as const;